Create the user modeling
I created the User model with email and password_digest attributes.

Email should be unique so I enforced it at database level with a unique
index and in the model validations.

Bcrypt gem was added to the Gemfile to be able to use
"has_secure_password" https://guides.rubyonrails.org/active_model_basics.html#securepassword